Rename MapPropertyFactory.CodeGenerator -> MapProperty
MapPropertyFactory is now MapProperty.Factory
